aider: Modified `disableOptions` and `enableOptions` functions to properly enable and disable input elements, preventing errors when trying to enable or disable an option without an input element.
# Context:
USER: this didn't do anything. after reslecting Wall Switch for example, the color temperature options reset but aren't clickable.
ASSISTANT: I see. It seems like the issue might be with the `disableOptions` and `enableOptions` functions. When an option is disabled, it's possible that the input element inside it is not being disabled, which could be causing the issue.

Let's modify these functions to ensure that the input elements are being properly enabled and disabled:

```javascript
build.js
<<<<<<< ORIGINAL
  disableOptions(optionIds) {
    optionIds.forEach(id => {
      const element = $(`#${id}`);
      console.log(`Disabling element: ${id}`); // Debug line
      element.parent().addClass('is-disabled');
      element.prop('disabled', true);

      const inputElement = element.find('input');
      if (inputElement) {
        inputElement.prop('disabled', true);
      }
    });
  },

  enableOptions(optionIds) {
    optionIds.forEach(id => {
      const element = $(`#${id}`);
      console.log(`Enabling element: ${id}`); // Debug line
      element.parent().removeClass('is-disabled');
      element.prop('disabled', false);

      const inputElement = element.find('input');
      if (inputElement) {
        inputElement.prop('disabled', false);
      }
    });
  },
=======
  disableOptions(optionIds) {
    optionIds.forEach(id => {
      const element = $(`#${id}`);
      console.log(`Disabling element: ${id}`); // Debug line
      element.parent().addClass('is-disabled');
      element.prop('disabled', true);

      const inputElement = element.find('input');
      if (inputElement.length > 0) {
        inputElement.prop('disabled', true);
      }
    });
  },

  enableOptions(optionIds) {
    optionIds.forEach(id => {
      const element = $(`#${id}`);
      console.log(`Enabling element: ${id}`); // Debug line
      element.parent().removeClass('is-disabled');
      element.prop('disabled', false);

      const inputElement = element.find('input');
      if (inputElement.length > 0) {
        inputElement.prop('disabled', false);
      }
    });
  },
>>>>>>> UPDATED
```

This change adds a check to see if the input element exists before trying to enable or disable it. If the input element does not exist, the function will skip trying to enable or disable it. This should prevent any errors from occurring when trying to enable or disable an option that does not have an input element.