aider: Added debug lines to the `enableOptions` function to log information about the elements being enabled.
# Context:
USER: it's starting to come together now. However, the `enableOptions` isn't working
ASSISTANT: From the code you've provided, it seems like the `enableOptions` function is trying to enable an option by removing the 'is-disabled' class from the parent of the element with the given ID and setting the 'disabled' property of the element to false.

If this function isn't working as expected, it could be due to a few reasons:

1. The element with the given ID doesn't exist or isn't found.
2. The parent of the element doesn't have the 'is-disabled' class.
3. The 'disabled' property of the element isn't being set correctly.

To help diagnose the issue, we can add more logging to the `enableOptions` function to trace its execution and see if it's being called as expected and if it's receiving the correct arguments.

Here's how we can modify the `enableOptions` function:

```javascript
build.js
<<<<<<< ORIGINAL
  enableOptions(optionIds) {
    optionIds.forEach(id => {
      const element = $(`#${id}`);
      console.log(`Enabling element: ${id}`); // Debug line
      element.parent().removeClass('is-disabled');
      element.prop('disabled', false);
    });
  },
=======
  enableOptions(optionIds) {
    optionIds.forEach(id => {
      const element = $(`#${id}`);
      console.log(`Enabling element: ${id}`); // Debug line
      console.log(`Found element with ID ${id}:`, element); // Added debug line
      console.log(`Parent of element:`, element.parent()); // Added debug line
      element.parent().removeClass('is-disabled');
      element.prop('disabled', false);
    });
  },
>>>>>>> UPDATED
```

This will give us more information about what's happening inside the `enableOptions` function. After making these changes, please run the code again and share the output of the console logs.